refactor : F/E Category 클래스에서 UploadTemplate에 대한 dependency 제거
